Fix: resolve a set of NPE at kubevirt networking

+++ apps/kubevirt-networking/app/src/main/java/org/onosproject/kubevirtnetworking/impl/NetworkAttachmentDefinitionWatcher.java
@@ -258,7 +258,9 @@ public class NetworkAttachmentDefinitionWatcher {
             log.trace("Process NetworkAttachmentDefinition {} removal event from API server.",
                     name);
 
-            adminService.removeNetwork(name);
+            if (adminService.network(name) != null) {
+                adminService.removeNetwork(name);
+            }
         }
 
         private boolean isMaster() {
